Revert dev and test envs to standard Rails logger
Leave `LogStashLogger` only in production env.
