+++ config/environments/development.rb
@@ -1,6 +1,4 @@
 Rails.application.configure do
-  config.logstash.type = :file
-
   config.cache_classes = false
 
   config.eager_load = false
