+++ config/environments/test.rb
@@ -3,8 +3,6 @@ Rails.application.config.before_configuration do
 end
 
 Rails.application.configure do
-  config.logstash.type = :file
-
   config.cache_classes = true
 
   config.eager_load = false
